[Issue-66] Add minimum for custom fee
