+++ packages/extension-base/src/utils/bitcoin/utxo-management.ts
@@ -3,46 +3,12 @@
 
 import { BITCOIN_DECIMAL, BTC_DUST_AMOUNT } from '@subwallet/extension-base/constants';
 import { _BitcoinApi } from '@subwallet/extension-base/services/chain-service/types';
-import { BitcoinFeeInfo, BitcoinFeeRate, DetermineUtxosForSpendArgs, FeeOption, InsufficientFundsError, UtxoResponseItem } from '@subwallet/extension-base/types';
+import { DetermineUtxosForSpendArgs, InsufficientFundsError, UtxoResponseItem } from '@subwallet/extension-base/types';
 import { BitcoinAddressType } from '@subwallet/keyring/types';
-import { BtcSizeFeeEstimator, getBitcoinAddressInfo, validateBitcoinAddress } from '@subwallet/keyring/utils';
+import { getBitcoinAddressInfo, validateBitcoinAddress } from '@subwallet/keyring/utils';
 import BigN from 'bignumber.js';
 
-export const combineBitcoinFee = (feeInfo: BitcoinFeeInfo, feeOptions?: FeeOption, feeCustom?: BitcoinFeeRate): BitcoinFeeRate => {
-  if (feeOptions && feeOptions !== 'custom') {
-    return feeInfo.options?.[feeOptions];
-  } else if (feeOptions === 'custom' && feeCustom) {
-    return feeCustom;
-  } else {
-    return feeInfo.options?.[feeInfo.options.default];
-  }
-};
-
-// https://github.com/leather-wallet/extension/blob/dev/src/app/common/transactions/bitcoin/utils.ts
-export function getSpendableAmount ({ feeRate,
-  recipients,
-  sender,
-  utxos }: {
-  utxos: UtxoResponseItem[];
-  feeRate: number;
-  recipients: string[];
-  sender: string;
-}) {
-  const balance = utxos.map((utxo) => utxo.value).reduce((prevVal, curVal) => prevVal + curVal, 0);
-
-  const size = getSizeInfo({
-    inputLength: utxos.length,
-    recipients,
-    sender
-  });
-  const fee = Math.ceil(size.txVBytes * feeRate);
-  const bigNumberBalance = new BigN(balance);
-
-  return {
-    spendableAmount: BigN.max(0, bigNumberBalance.minus(fee)),
-    fee
-  };
-}
+import { getSizeInfo, getSpendableAmount } from './common';
 
 // https://github.com/leather-wallet/extension/blob/dev/src/app/common/transactions/bitcoin/utils.ts
 // Check if the spendable amount drops when adding a utxo. If it drops, don't use that utxo.
@@ -93,38 +59,6 @@ export function filterUneconomicalUtxos ({ feeRate,
       return spendableAmountWithout.gt(spendableAmount) ? utxosWithout : utxos;
     }
   }, [...filteredAndSortUtxos]).reverse();
-}
-
-// Source: https://github.com/leather-wallet/extension/blob/dev/src/app/common/transactions/bitcoin/utils.ts
-export function getSizeInfo (payload: {
-  inputLength: number;
-  recipients: string[];
-  sender: string;
-}) {
-  const { inputLength, recipients, sender } = payload;
-  const senderInfo = validateBitcoinAddress(sender) ? getBitcoinAddressInfo(sender) : null;
-  const inputAddressTypeWithFallback = senderInfo ? senderInfo.type : BitcoinAddressType.p2wpkh;
-  const outputMap: Record<string, number> = {};
-
-  for (const recipient of recipients) {
-    const recipientInfo = validateBitcoinAddress(recipient) ? getBitcoinAddressInfo(recipient) : null;
-    const outputAddressTypeWithFallback = recipientInfo ? recipientInfo.type : BitcoinAddressType.p2wpkh;
-    const outputKey = outputAddressTypeWithFallback + '_output_count';
-
-    if (outputMap[outputKey]) {
-      outputMap[outputKey]++;
-    } else {
-      outputMap[outputKey] = 1;
-    }
-  }
-
-  const txSizer = new BtcSizeFeeEstimator();
-
-  return txSizer.calcTxSize({
-    input_script: inputAddressTypeWithFallback,
-    input_count: inputLength,
-    ...outputMap
-  });
 }
 
 // https://github.com/leather-wallet/extension/blob/dev/src/app/common/transactions/bitcoin/coinselect/local-coin-selection.ts
